requirements.txt
grpcio==1.62.2
protobuf==4.25.3
fastapi[standard]
loguru